docs: 테스트용 코드를 다시 onboarding 패키지로 이동, 사용하지 않는 부분 삭제
