--- app/src/main/java/com/android/hanple/ui/test/LocalUser.kt
@@ -1,7 +1,0 @@
-package com.android.hanple.ui.test
-
-data class LocalUser(
-    val email: String,
-    val password: String,
-    val username: String
-)
