--- app/src/main/res/layout/activity_local_sign_up.xml
@@ -1,90 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.test.LocalSignUpActivity">
-
-    <ImageView
-        android:id="@+id/iv_back"
-        android:layout_width="20dp"
-        android:layout_height="20dp"
-        android:src="@drawable/ic_back_8x12dp"
-        android:layout_margin="32dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-
-
-
-    <LinearLayout
-        android:id="@+id/edit_texts_li"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/iv_back"
-        android:gravity="center_horizontal">
-        <TextView
-            android:id="@+id/title_tv"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/onboarding_welcome"
-            android:layout_marginBottom="20dp"
-            android:textAlignment="center"
-            android:textColor="@color/darkblue"
-            android:fontFamily="@font/pretendard_regular"
-            android:textSize="20sp"
-            android:textStyle="bold" />
-        <EditText
-            android:id="@+id/et_email"
-            style="@style/etGeneral"
-            android:inputType="textEmailAddress"
-            android:fontFamily="@font/pretendard_regular"
-            android:textSize="18sp"
-            android:hint="@string/email" />
-        <EditText
-            android:id="@+id/et_password"
-            style="@style/etGeneral"
-            android:layout_marginTop="12dp"
-            android:inputType="textPassword"
-            android:fontFamily="@font/pretendard_regular"
-            android:textSize="18sp"
-            android:hint="@string/password" />
-        <EditText
-            android:id="@+id/et_username"
-            style="@style/etGeneral"
-            android:layout_marginTop="12dp"
-            android:fontFamily="@font/pretendard_regular"
-            android:textSize="18sp"
-            android:hint="@string/nickname" />
-        <TextView
-            android:id="@+id/tv_error"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="이메일을 작성해 주세요."
-            android:textColor="@color/tv_is_valid_selector"
-            android:fontFamily="@font/pretendard_regular"
-            android:textSize="14sp"
-            android:layout_marginTop="12dp"
-            android:layout_marginStart="30dp"
-            android:layout_marginEnd="30dp"
-            android:maxLines="1"/>
-    </LinearLayout>
-
-    <androidx.appcompat.widget.AppCompatButton
-        android:id="@+id/btn_sign_up"
-        style="@style/btnGeneral1st"
-        android:text="@string/onboarding_sign_up"
-        android:fontFamily="@font/pretendard_regular"
-        android:textSize="20sp"
-        android:layout_marginTop="12dp"
-        app:layout_constraintTop_toBottomOf="@id/edit_texts_li"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"/>
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
